+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed comments staying in the buffer for too long (see #117).
 - Fixed missing scopped variables in suggestions.
 - Fixed methodmaps' property parsing bug.
+- Improved include's path resolution, there should be less collisions if you have two files with the same name in the same project workspace.
 - Improved BUG issues templates (see #116).
 - Improved JSDoc completions provider.