+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Release Notes
 
-## [2.3.3]
+## [2.4.0]
 
 ### Added